+++ aqt/installer.py
@@ -20,14 +20,12 @@
 # IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 # CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-import concurrent.futures
 import functools
+import multiprocessing.dummy
 import os
 import subprocess
-import time
 from logging import getLogger
 from operator import and_
-from time import sleep
 
 import requests
 
@@ -59,6 +57,7 @@ class QtInstaller:
     def retrieve_archive(self, package):
         archive = package.archive
         url = package.url
+        self.logger.info("Downloading {}...".format(url))
         try:
             r = requests.get(url, allow_redirects=False, stream=True)
             if r.status_code == 302:
@@ -68,7 +67,7 @@ class QtInstaller:
                 r = requests.get(newurl, stream=True)
         except requests.exceptions.ConnectionError as e:
             self.logger.warning("Caught download error: %s" % e.args)
-            return None
+            return False
         else:
             try:
                 with open(archive, 'wb') as fd:
@@ -76,21 +75,31 @@ class QtInstaller:
                         fd.write(chunk)
             except Exception as e:
                 self.logger.warning("Caught download error: %s" % e.args)
-                return None
-        return archive
+                return False
+        return True
 
-    def extract_archive(self, archive):
-        py7zr.SevenZipFile(archive).extractall(path=self.base_dir)
-        os.unlink(archive)
-        return archive, time.process_time()
-
-    def extract_archive_ext(self, archive):
-        if self.base_dir is not None:
-            subprocess.run([self.command, 'x', '-aoa', '-bd', '-y', '-o{}'.format(self.base_dir), archive])
+    def extract_archive(self, package):
+        try:
+            archive = package.archive
+            py7zr.SevenZipFile(archive).extractall(path=self.base_dir)
+            os.unlink(archive)
+        except Exception:
+            return False
         else:
-            subprocess.run([self.command, 'x', '-aoa', '-bd', '-y', archive])
-        os.unlink(archive)
-        return archive, 0
+            return True
+
+    def extract_archive_ext(self, package):
+        try:
+            archive = package.archive
+            if self.base_dir is not None:
+                subprocess.run([self.command, 'x', '-aoa', '-bd', '-y', '-o{}'.format(self.base_dir), archive])
+            else:
+                subprocess.run([self.command, 'x', '-aoa', '-bd', '-y', archive])
+            os.unlink(archive)
+        except Exception:
+            return False
+        else:
+            return True
 
     def install(self):
         qt_version, target, arch = self.qt_archives.get_target_config()
@@ -99,66 +108,33 @@ class QtInstaller:
         else:
             extractor = self.extract_archive_ext
         archives = self.qt_archives.get_archives()
+        p = multiprocessing.dummy.Pool(3)
+        ret_arr = p.map(self.retrieve_archive, archives)
+        ret = functools.reduce(and_, ret_arr)
+        if ret:
+            self.logger.info("Downloads are Completed.")
+        else:
+            self.logger.error("Failed to download.")
+        p = multiprocessing.dummy.Pool()
+        ret_arr = p.map(extractor, archives)
+        ret = functools.reduce(and_, ret_arr)
+        if ret:
+            self.logger.info("Extractions are Completed.")
+        else:
+            self.logger.error("Failed to extract archvies.")
 
-        # retrieve files from download site
-        with concurrent.futures.ProcessPoolExecutor() as pexec:
-            download_task = []
-            completed_downloads = []
-            extract_task = []
-            completed_extract = []
-            with concurrent.futures.ThreadPoolExecutor() as texec:
-                for ar in archives:
-                    self.logger.info("Downloading {}...".format(ar.url))
-                    download_task.append(texec.submit(self.retrieve_archive, ar))
-                    completed_downloads.append(False)
-                    completed_extract.append(False)
-                while True:
-                    for i, t in enumerate(download_task):
-                        if completed_downloads[i]:
-                            if not completed_extract[i] and i < len(extract_task) and extract_task[i].done():
-                                (archive, elapsed) = extract_task[i].result()
-                                self.logger.info("Done {} extraction in {:.8f}.".format(archive, elapsed))
-                                completed_extract[i] = True
-                        elif t.done():
-                            archive = t.result()
-                            if archive is None:
-                                self.logger.error("Failed to download.")
-                                exit(1)
-                            self.logger.info("Extracting {}...".format(archive))
-                            extract_task.append(pexec.submit(extractor, archive))
-                            completed_downloads[i] = True
-                    if functools.reduce(and_, completed_downloads):
-                        self.logger.info("Downloads are Completed.")
-                        break
-                    else:
-                        for i, t in enumerate(extract_task):
-                            if not completed_extract[i] and t.done():
-                                (archive, elapsed) = t.result()
-                                self.logger.info("Done {} extraction in {:.8f}.".format(archive, elapsed))
-                                completed_extract[i] = True
-                        sleep(0.05)
-            while True:
-                for i, t in enumerate(extract_task):
-                    if not completed_extract[i] and t.done():
-                        (archive, elapsed) = t.result()
-                        self.logger.info("Done {} extraction in {:.8f}.".format(archive, elapsed))
-                        completed_extract[i] = True
-                if functools.reduce(and_, completed_extract):
-                    break
-                else:
-                    sleep(0.5)
-            # finalize
-            if qt_version != "Tools":  # tools installation
-                if arch.startswith('win64_mingw'):
-                    arch_dir = arch[6:] + '_64'
-                elif arch.startswith('win32_mingw'):
-                    arch_dir = arch[6:] + '_32'
-                elif arch.startswith('win'):
-                    arch_dir = arch[6:]
-                else:
-                    arch_dir = arch
-                self.make_conf_files(qt_version, arch_dir)
-            self.logger.info("Finished installation")
+        # finalize
+        if qt_version != "Tools":  # tools installation
+            if arch.startswith('win64_mingw'):
+                arch_dir = arch[6:] + '_64'
+            elif arch.startswith('win32_mingw'):
+                arch_dir = arch[6:] + '_32'
+            elif arch.startswith('win'):
+                arch_dir = arch[6:]
+            else:
+                arch_dir = arch
+            self.make_conf_files(qt_version, arch_dir)
+        self.logger.info("Finished installation")
 
     def make_conf_files(self, qt_version, arch_dir):
         """Make Qt configuration files, qt.conf and qtconfig.pri"""